requirements.txt
asgiref==3.9.1
Django==4.2.23
Pillow==11.3.0
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
gunicorn==23.0.0
packaging==25.0